ボタン長押しによるマルチシーケンス切り替え機能を追加
- ボタン0の長押し（1秒、停止中のみ）でシーケンスを循環切り替え
- シーケンスデータを sequence0/1/2 に分割、PROGMEM ポインタ配列で管理
- シーケンスごとのメタデータ（ループ設定・ステップ数等）を配列化
- types.h: Step 構造体の step/next/loopNext を uint16_t に拡張（256以上のステップに対応）
- resetSequenceState() 追加（シーケンス切り替え時に全状態をリセット）

+++ mgma2_auto_sevseg_button/types.h
@@ -49,17 +49,17 @@ struct ButtonState {
 
 // --- ステップ定義 ---
 struct Step {
-  byte step;
+  uint16_t step;
   uint8_t hat;
   uint16_t button;
   bool press;
   uint16_t _dur;
-  byte next;
+  uint16_t next;
   byte loop;
-  byte loopNext;
+  uint16_t loopNext;
 
   uint32_t duration() {
-    return _dur * 100;
+    return (uint32_t)_dur * 100;
   }
 };
